requirements.txt
# ⛏️ VideoMine - Dependencies
flask
flask-cors
jinja2
yt-dlp
openai-whisper
anthropic
